+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+## Bug fixes
+- Fix armada installation for docker >= 18.09.0.
+
 ## 2.11.0 (2021-02-02)
 
 We do best effort to support docker versions 1.12.0 - 17.12.1 with this release.